Add files via upload
The file contains steps to connect to Twitters streaming API and extract relevant tweets. In addition, it also contains steps undertaken to structure the data (originally in JSON) and analyze it in greater detail. Finally, it includes the steps carried out to build a ML model that can accurately classify the sentiment of a given tweet.